refactor: correct file structure
