+++ web/src/app/api/workflowData/asyncUpdateWorkflow.ts
@@ -1,5 +1,5 @@
 import { serverTimestamp, updateDoc, doc } from 'firebase/firestore';
-import { db } from '@/app/lib/firebase';
+import { db } from '@/lib/firebase';
 
 const asyncUpdateWorkflow = async (uid: string, id = '', payload = {}) => {
   const workflowRef = doc(db, 'users', uid, 'scripts', id);
