+++ web/src/components/landingPage/mainPicture.png
