+++ web/src/nodes/EnterSubmitEvent.tsx
@@ -1,35 +1,40 @@
-import React, { useState } from 'react';
+import { useState, ChangeEvent } from 'react';
 import { Handle, Position } from 'reactflow';
 
-import { MdAdsClick } from 'react-icons/md';
+import { LuSend } from 'react-icons/lu';
 
-import useStore from '../store';
+import useStore from '@/store';
 import { shallow } from 'zustand/shallow';
 import ContextMenu from './utils/ContextMenu';
 
-type inputEvent = React.ChangeEvent<HTMLInputElement>;
+type inputEvent = ChangeEvent<HTMLInputElement>;
+type textareaEvent = ChangeEvent<HTMLTextAreaElement>;
 
 const selector = (id: string) => (store: Store) => ({
   setXPath: (e: inputEvent) => store.updateNode(id, { XPath: e.target.value }),
   setDescription: (e: inputEvent) => store.updateNode(id, { description: e.target.value }),
+  setValue: (e: textareaEvent) => store.updateNode(id, { value: e.target.value }),
 });
 
-type clickObj = {
-  label: 'Click';
+type inputObj = {
+  label: 'Input';
   disable: boolean;
+  value: string;
   XPath: string;
   description: string;
+  inputType: string;
+  isConnectable: boolean;
+  maxConnections: number;
 };
 
 interface ClickEventProps {
   id: string;
-  isConnectable: boolean;
-  data: clickObj;
+  data: inputObj;
 }
 
-function ClickEvent({ id, isConnectable, data }: ClickEventProps) {
+function EnterSubmitEvent({ id, data }: ClickEventProps) {
   const { setXPath, setDescription } = useStore(selector(id), shallow);
-
+  const { isConnectable } = data;
   const [stashXpath, setStashXpath] = useState(data.XPath);
   const [stashDescription, setStashDescription] = useState(data.description);
 
@@ -42,47 +47,48 @@ function ClickEvent({ id, isConnectable, data }: ClickEventProps) {
   };
 
   return (
-    <ContextMenu id={id} color={'text-customGreen-400'}>
+    <ContextMenu id={id} color={'text-customDarkBlue-400'}>
       <div
-        className={`overflow-hidde w-[228px] overflow-hidden rounded-nodebase border border-customGreen-400 bg-white ${
+        className={` w-[228px] overflow-hidden rounded-nodebase border border-customDarkBlue-400 bg-white ${
           data.disable ? 'toggleOpacity' : ''
-        } `}
+        }`}
       >
         <Handle
           type='target'
           position={Position.Left}
           isConnectable={isConnectable}
-          className=' react-flow__handle-target after:border after:border-customGreen-400'
+          className=' react-flow__handle-target after:border after:border-customDarkBlue-400'
           style={{ left: '-8px', top: '81px' }}
         />
-        <div className=' flex h-[57px] items-center justify-center border border-b-customGreen-400 bg-customGreen-50'>
+        <div className=' flex h-[56px] items-center justify-center border border-b-customDarkBlue-400 bg-customDarkBlue-50'>
           <div className=' flex items-center gap-2'>
-            <span className=' text-customGreen-400'>
-              <MdAdsClick size='24px' />
+            <span className=' text-customDarkBlue-400'>
+              <LuSend size='22px' />
             </span>
-            <div className=' text-2xl font-medium text-customGreen-400'>{data.label}</div>
+            <div className=' text-2xl font-medium text-customDarkBlue-400'>{data.label}</div>
           </div>
         </div>
-        <div className=' pm-5 flex flex-col pb-5 pl-5 pr-5'>
+        <div className=' flex flex-col pb-5 pl-5 pr-5'>
           <input
-            id='clickDescription'
-            name='clickDescription'
+            id='textDescription'
+            name='textDescription'
             onChange={descriptionChange}
             value={stashDescription}
             onBlur={setDescription}
-            placeholder='click for.....'
             maxLength={21}
+            placeholder='submit for.....'
             className='nodrag mt-3 rounded-sm text-center text-gray-400 outline-none focus:underline'
           />
-          <label htmlFor='text' className=' mb-1 text-customGreen-500'>
+          <label htmlFor='text' className=' mb-1 text-customDarkBlue-500'>
             XPath：
           </label>
           <input
-            type='text'
+            id='text'
+            name='text'
             value={stashXpath}
             onChange={xpathChange}
             onBlur={setXPath}
-            className='nodrag rounded-md border border-customGreen-500 p-1 outline-none'
+            className='nodrag rounded-md border border-customDarkBlue-500 p-1 outline-none'
             placeholder='xpath.....'
           />
         </div>
@@ -91,7 +97,7 @@ function ClickEvent({ id, isConnectable, data }: ClickEventProps) {
           position={Position.Right}
           id='click'
           isConnectable={isConnectable}
-          className=' react-flow__handle-source after:border after:border-customGreen-400'
+          className=' react-flow__handle-source after:border after:border-customDarkBlue-400'
           style={{ right: '-8px', top: '140px' }}
         />
       </div>
@@ -99,4 +105,4 @@ function ClickEvent({ id, isConnectable, data }: ClickEventProps) {
   );
 }
 
-export default ClickEvent;
+export default EnterSubmitEvent;
