+++ web/types/global.d.ts
@@ -1,14 +1,10 @@
-import { Edge, Node } from 'reactflow';
+import { Edge, Node, XYPosition } from 'reactflow';
 
 declare global {
   type Flow = {
     nodes: Node[];
     edges: Edge[];
-    viewport: {
-      x: number;
-      y: number;
-      zoom: number;
-    };
+    viewport: XYPosition;
   };
 
   type Store = {
